Fix YumeCharacter softlock
YumeCharacter uses tweening to move characters. To make sure characters
cant do anything while tweening theres a special variable called
is_busy. is_busy is supposed to get reset back to false after a
character finishes moving. However when a level is changed in the middle
of character movement, the end of the tween never reaches.

To fix this, during the tween if the level transitions, the tween skips
to its end point. Which allows await tween.finished to run and reset
is_busy